+++ projects/volt/src/lib/components/listbox/index.ts
@@ -3,3 +3,4 @@ export * from './listbox-option';
 export * from './listbox-section';
 export * from './listbox-header';
 export * from './listbox-trigger';
+export { listboxOptionVariants, type ListboxOptionVariants } from './variants';